+++ packages/react/src/index.ts
@@ -65,4 +65,5 @@ export * from "./display";
 export * from "./map";
 export * from "./preview";
 export * from "./previewFixtures";
+export * from "./saveSlots";
 export * from "./cameraShake";